Add Request Ajax Relates #25
