+++ src/router.js
@@ -1,5 +1,8 @@
 const fs = require('fs');
 const path = require('path');
+const queryString = require('querystring');
+
+const wordFilter = require('./utils');
 
 const extentions = {
   '.css': 'text/css',
@@ -7,24 +10,6 @@ const extentions = {
   '.html': 'text/html',
   '.png': 'image/png',
 };
-
-// const pathFile = path.join(__dirname, 'words.txt');
-// const pathFileJSON = path.join(__dirname, 'dictionary.json');
-// const arr = [];
-
-// fs.readFile(pathFile, 'utf8', (err, data) => {
-//   if (err) {
-//     throw err;
-//   }
-//   const myData = data.split('\n');
-//   // myData.forEach((item) => arr.push(item));
-
-//   fs.writeFile(pathFileJSON, JSON.stringify(myData, null, 2), (err) => {
-//     if (err) {
-//       throw err;
-//     }
-//   });
-// });
 
 const router = (request, response) => {
   const endPoint = request.url;
