+++ src/utils.js
@@ -1,0 +1,6 @@
+const wordFilter = (frag, array) => array.filter(
+  (word) => word.toLowerCase().trim().substring(0, frag.length)
+      === frag.toLowerCase().trim(),
+);
+
+module.exports = wordFilter;
